Pre-paint: Handle missing tree builder context correctly.
The tree builder is ever only missing if DCHECKs are disabled, which
explains why this bug wasn't reproducible with DHCECKs enabled.

In this mode, we'd never increment inner_fragmentainer_idx, which would
cause us to discard all existing FragmentData entries, so that we'd have
more NGPhysicalBoxFragment objects than FragmentData objects, and
LayoutBox::FragmentDataFromPhysicalFragment() would crash.

Structure the code in such a way that we'll hopefully not make a similar
mistake again.

Bug: 1278652
Change-Id: I3ec35eb81b330f5a40f9b2469218eaecff7dc586