[Anonymous iframe] Pass whether a window should commit anonymously.
A window is anonymous means a window is loaded under an anonymous
iframe which do not share credentials and only aim to load public data.

This attribute is not currently enabled on initial empty document, see:
https://docs.google.com/document/d/1JcBLdl7P6ZrtIT63D-dsB52yxXOuVrHS2gamhU0wZwk/edit#heading=h.oyyrxad04wu6

This CL exposed anonymous to window and renderer process, which aimed
to be used in a follow-up CL to disable password autofill.
For further info, please refer to:
https://docs.google.com/document/d/1UHEC10kjxfgNmt-ejGkuxgvv63_-dsEH-DhVF_Sthak/edit#

Bug: 1233858
Change-Id: If63a95c3306fb31c62fcb01aab6e13745f025b3d