[LayoutNG] OOF in nested multicol with spanner CB
If an OOF had a spanner CB (or a CB inside a spanner) inside a nested
multicol, and the inner multicol had no content, we would crash because
we expected the multicol to have a column set.

However, if the spanner is the containing block (or the CB is inside
a spanner), we shouldn't lay out the OOF inside the inner multicol, but
in the outer multicol. This CL ensures that OOFs in this case are
properly propagated to the next multicol so that they can be laid out
there instead.

To accomplish this, a new bool was added to the ContainingBlock
stored on the OutOfFlowPositionedNode so that we can tell when the
containing block is inside a column spanner.

Bug: 1278277
Change-Id: I6fe6d9f499cd9622cd63a99178071b17aaea3dee