[css-pseudo] Add support for text-decoration in highlight pseudos
This patch adds support for adding text-decoration on highlight pseudos
::highlight and ::target-text.

To do this, it exposes HighlightPaintingUtils::HighlightPseudoStyle()
so we can use it from NGHighlightPainter::Paint().
There we use NGTextDecorationPainter to paint the text decorations
defined by the highlight pseudo.

To make the tests pass when HighlightInheritance is enabled,
we had to stop propagating text decorations
in StyleAdjuster::AdjustComputedStyle() for the highlight pseudos.

BUG=1274174

TEST=external/wpt/css/css-highlight-api/painting/custom-highlight-painting-text-decoration-001.html
TEST=external/wpt/css/css-highlight-api/painting/custom-highlight-painting-text-decoration-dynamic-001.html
TEST=external/wpt/css/css-pseudo/target-text-text-decoration-001.html

Change-Id: I9731fd7dbef938a8470d190310d20612d95e72b8